added builderlayout to scene-builder cuz no layout
